+++ benchmarks/js-framework-benchmark/src/App.compiled.js
@@ -1,14 +1,14 @@
 
-import { createSignal, createEffect } from '/runtime/core.js';
-import { mountPrimitives as dom_mountPrimitives, walk } from '/runtime/dom.js';
-import { List } from '/runtime/primitives/list.js';
+import { createSignal, createEffect } from 'pulse/runtime';
+import { mountPrimitives as dom_mountPrimitives, walk } from 'pulse/runtime/dom';
+import { List } from 'pulse/runtime/list';
 
 
 
 
 // Static Template
 const _tpl = document.createElement('template');
-_tpl.innerHTML = `<div class="data-v-1eyp pulse-component-app"><div class="container"><div class="jumbotron"><div class="row"><div class="col-md-6"><h1>Pulse ("keyed")</h1></div><div class="col-md-6"><div class="row"><div class="col-sm-6 smallpad"><button type="button" class="btn btn-primary btn-block" id="run" data-on-click="{run}">Create 1,000 rows</button></div><div class="col-sm-6 smallpad"><button type="button" class="btn btn-primary btn-block" id="runlots" data-on-click="{runLots}">Create 10,000 rows</button></div><div class="col-sm-6 smallpad"><button type="button" class="btn btn-primary btn-block" id="add" data-on-click="{add}">Append 1,000 rows</button></div><div class="col-sm-6 smallpad"><button type="button" class="btn btn-primary btn-block" id="update" data-on-click="{update}">Update every 10th row</button></div><div class="col-sm-6 smallpad"><button type="button" class="btn btn-primary btn-block" id="clear" data-on-click="{clear}">Clear</button></div><div class="col-sm-6 smallpad"><button type="button" class="btn btn-primary btn-block" id="swaprows" data-on-click="{swapRows}">Swap Rows</button></div></div></div></div></div><table class="table table-hover table-striped test-data"><tbody><pulse-list each="{state.data}" as="row" key="{row.id}" data-bindings="[{&quot;type&quot;:&quot;text&quot;,&quot;path&quot;:[0,0,0],&quot;expr&quot;:&quot;row.id&quot;},{&quot;type&quot;:&quot;text&quot;,&quot;path&quot;:[0,1,0,0],&quot;expr&quot;:&quot;row.label&quot;}]" data-template-id="tmpl_0" style="display:contents"></pulse-list></tbody></table><span class="preloadicon glyphicon glyphicon-remove" aria-hidden="true"></span></div></div>`; // Escape backticks
+_tpl.innerHTML = `<div class="data-v-1eyp pulse-component-app"><div class="container"><div class="jumbotron"><div class="row"><div class="col-md-6"><h1>Pulse ("keyed")</h1></div><div class="col-md-6"><div class="row"><div class="col-sm-6 smallpad"><button type="button" class="btn btn-primary btn-block" id="run" data-on-click="{run}">Create 1,000 rows</button></div><div class="col-sm-6 smallpad"><button type="button" class="btn btn-primary btn-block" id="runlots" data-on-click="{runLots}">Create 10,000 rows</button></div><div class="col-sm-6 smallpad"><button type="button" class="btn btn-primary btn-block" id="add" data-on-click="{add}">Append 1,000 rows</button></div><div class="col-sm-6 smallpad"><button type="button" class="btn btn-primary btn-block" id="update" data-on-click="{update}">Update every 10th row</button></div><div class="col-sm-6 smallpad"><button type="button" class="btn btn-primary btn-block" id="clear" data-on-click="{clear}">Clear</button></div><div class="col-sm-6 smallpad"><button type="button" class="btn btn-primary btn-block" id="swaprows" data-on-click="{swapRows}">Swap Rows</button></div></div></div></div></div><table class="table table-hover table-striped test-data"><tbody><pulse-list each="{state.data}" as="row" key="{row.id}" data-bindings="[{&quot;type&quot;:&quot;attribute&quot;,&quot;path&quot;:[0],&quot;name&quot;:&quot;class&quot;,&quot;expr&quot;:&quot;state.selected === row.id ? \&quot;danger\&quot; : \&quot;\&quot;&quot;},{&quot;type&quot;:&quot;text&quot;,&quot;path&quot;:[0,0,0],&quot;expr&quot;:&quot;row.id&quot;},{&quot;type&quot;:&quot;text&quot;,&quot;path&quot;:[0,1,0,0],&quot;expr&quot;:&quot;row.label&quot;}]" data-template-id="tmpl_0" style="display:contents"></pulse-list></tbody></table><span class="preloadicon glyphicon glyphicon-remove" aria-hidden="true"></span></div></div>`; // Escape backticks
 
 export default function App(props) {
   props = props || {};
@@ -46,12 +46,12 @@ export default function App(props) {
   function buildData(count) {
   const data = new Array(count);
   for (let i = 0; i < count; i++) {
-    get_data[i] = {
+    data[i] = {
       id: nextId++,
       label: `${adjectives[random(adjectives.length)]} ${colors[random(colors.length)]} ${nouns[random(nouns.length)]}`,
     };
   }
-  return get_data;
+  return data;
 }
   function run() {
   state.data = buildData(1000);
@@ -117,6 +117,7 @@ export default function App(props) {
     nouns: nouns, 
     nextId: nextId, 
     props: props, 
+    state: state, 
   }; 
 
   const mountPrimitives = (cont) => {
@@ -127,11 +128,6 @@ export default function App(props) {
        createEffect: createEffect, 
     }, {"tmpl_0":"<tr><td class=\"col-md-1\"> </td><td class=\"col-md-4\"><a data-on-click=\"{() => select(row.id)}\"> </a></td><td class=\"col-md-1\"><a data-on-click=\"{() => remove(row.id)}\"><span class=\"glyphicon glyphicon-remove\" aria-hidden=\"true\"></span></a></td><td class=\"col-md-6\"></td></tr>"}); 
   }; 
-
-  createEffect(() => {
-    const el = walk(container, [0]);
-    if (el) el.setAttribute('class', state.selected === row.id ? "danger" : "");
-  });
 
   const handlers = { random: random, buildData: buildData, run: run, runLots: runLots, add: add, update: update, clear: clear, swapRows: swapRows, remove: remove, select: select };
   container.__pulseHandlers = handlers;
@@ -152,5 +148,6 @@ export default function App(props) {
   }
 
   mountPrimitives(container);
+  if (container && container.setAttribute) container.setAttribute('data-p-h', '1');
   return container;
 }